Handle error situations from the API better
Currently it's possible to break the entire thing if you enter
characters into the TTS API which cannot be handled by the API. In these
situations the API will just send an empty HTML page, which will be
cached by the backend and sent to all clients as an MP3 file which
breaks the audio playback in the browser.

To avoid this, add some better error handling in the MP3 file retrieval
and do not create file system write streams until we are absolutely
certain that the API returned an MP3 file for us.